requirements.txt
python-dotenv==0.18.0
PyYAML==5.4.1
dnspython==1.16.0
urllib3==1.26.5
pymongo==3.11.4
